+++ Countly.m
@@ -298,6 +298,7 @@ static dispatch_once_t onceToken;
     CountlyContentBuilderInternal.sharedInstance.enableContentReloadOnStall = config.content.getEnableContentReloadOnStall;
     CountlyContentBuilderInternal.sharedInstance.contentReloadOnStallTimeout = config.content.getContentReloadOnStallTimeout / 1000.0;
     CountlyContentBuilderInternal.sharedInstance.disableZoom = config.content.getDisableZoom;
+    CountlyContentBuilderInternal.sharedInstance.disableRotation = config.content.getDisableRotation;
     CountlyContentBuilderInternal.sharedInstance.contentURLHandler = config.content.getContentURLHandler;
 #endif